Replaced in tests/ double with float, with corresponding matrix types CV_32

+++ tests/performance/auxiliaries/RecorderPerformance/main.cpp
@@ -57,7 +57,7 @@ void record()
 {
 
   int sz[] = {40, 40, 40};
-  cv::Mat mat1(3, sz, CV_64F, cv::Scalar::all(0));
+  cv::Mat mat1(3, sz, CV_32F, cv::Scalar::all(0));
   cv::Mat mat2(3, sz, CV_8S, cv::Scalar::all(-42));
   cv::Mat mat3(100,100,CV_8UC3);
